requirements.txt
numpy
pandas
scipy
scikit-learn
matplotlib
tqdm
pyyaml
soundfile
librosa
torch
torchvision
torchaudio
webrtcvad
noisereduce
opensmile
shap